# backend/docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearer-key":[]}],"tags":[{"name":"Clients","description":"Client management and churn prediction endpoints"},{"name":"Statistics","description":"Provides global business metrics and aggregated client data."},{"name":"Authentication","description":"Endpoints for authentication and JWT token generation"}],"paths":{"/clients":{"get":{"tags":["Clients"],"summary":"List active clients","operationId":"getClientList","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Missing or invalid JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"400":{"description":"Invalid pagination or sorting parameters","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"200":{"description":"Clients retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageClientListResponseDTO"}}}}}},"put":{"tags":["Clients"],"summary":"Update client","operationId":"updateClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientUpdateRequestDTO"}}},"required":true},"responses":{"404":{"description":"Client not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"401":{"description":"Missing or invalid JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"200":{"description":"Client updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientResponseDTO"}}}}}},"post":{"tags":["Clients"],"summary":"Create client","operationId":"createClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientCreateRequestDTO"}}},"required":true},"responses":{"201":{"description":"Client created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientResponseDTO"}}}},"401":{"description":"Missing or invalid JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}}}}},"/login":{"post":{"tags":["Authentication"],"summary":"User login","description":"Authenticates a user using email and password and returns a JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"200":{"description":"Authentication successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/clients/{id}":{"get":{"tags":["Clients"],"summary":"Get client by ID","operationId":"getClientById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"401":{"description":"Missing or invalid JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"200":{"description":"Client found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientResponseDTO"}}}},"400":{"description":"Invalid ID format provided","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"404":{"description":"Client not found with the given ID","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}}}},"delete":{"tags":["Clients"],"summary":"Delete client","operationId":"deleteClient","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Client not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"401":{"description":"Missing or invalid JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"204":{"description":"Client deactivated successfully"}}}},"/clients/prediction/{dni}":{"get":{"tags":["Clients"],"summary":"Get churn prediction for client by DNI","operationId":"getClientPredictionByDni","parameters":[{"name":"dni","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Invalid DNI or prediction error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"404":{"description":"Client not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"401":{"description":"Missing or invalid JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"200":{"description":"Prediction generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientFullResponseDTO"}}}}}}},"/clients/clients/statistics/{id}":{"get":{"tags":["Clients"],"summary":"Get statistics for client by ID","operationId":"getClientStatistics","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Client not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}},"200":{"description":"Statistics generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClientStatisticsDTO"}}}},"401":{"description":"Missing or invalid JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}}}}},"/api/stats":{"get":{"tags":["Statistics"],"summary":"Get statistics","operationId":"getGlobalStats","responses":{"200":{"description":"Statistics retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GlobalStatisticsDTO"}}}},"401":{"description":"Missing or invalid JWT token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorStatusResponseDTO"}}}}}}},"/api/stats/global":{"get":{"tags":["global-stats-controller"],"operationId":"getGlobalStats_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}},"components":{"schemas":{"ClientUpdateRequestDTO":{"required":["id"],"type":"object","properties":{"id":{"minimum":1,"type":"integer","format":"int64"},"clientName":{"maxLength":100,"minLength":3,"type":"string"},"clientPhone":{"type":"string"},"nearLocation":{"type":"boolean"},"birthDate":{"type":"string","format":"date"}}},"ErrorStatusResponseDTO":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"status":{"type":"integer","format":"int32"},"desc":{"type":"string"},"message":{"type":"string"},"fields":{"type":"array","items":{"$ref":"#/components/schemas/ErrorStatusResponseFieldDTO"}}}},"ErrorStatusResponseFieldDTO":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}},"ClientResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"clientName":{"type":"string"},"active":{"type":"boolean"},"gender":{"type":"string","enum":["MALE","FEMALE"]},"clientPhone":{"type":"string"},"nearLocation":{"type":"boolean"},"age":{"type":"integer","format":"int32"}}},"LoginRequestDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"TokenDTO":{"type":"object","properties":{"token":{"type":"string"}}},"ClientCreateRequestDTO":{"required":["active","birthDate","clientName","clientPhone","contractPeriod","dni","gender","groupVisits","nearLocation","promoFriends"],"type":"object","properties":{"clientName":{"maxLength":100,"minLength":3,"type":"string"},"dni":{"type":"string"},"gender":{"type":"string","enum":["MALE","FEMALE"]},"birthDate":{"type":"string","format":"date"},"clientPhone":{"type":"string"},"nearLocation":{"type":"boolean"},"partnerId":{"type":"integer","format":"int64"},"promoFriends":{"type":"boolean"},"contractPeriod":{"maximum":12,"minimum":1,"type":"integer","format":"int32"},"groupVisits":{"type":"boolean"},"active":{"type":"boolean"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"ClientListResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"clientName":{"type":"string"},"active":{"type":"boolean"},"gender":{"type":"string","enum":["MALE","FEMALE"]},"clientPhone":{"type":"string"},"nearLocation":{"type":"boolean"},"age":{"type":"integer","format":"int32"}}},"PageClientListResponseDTO":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ClientListResponseDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"ClientFullResponseDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"clientName":{"type":"string"},"clientPhone":{"type":"string"},"age":{"type":"integer","format":"int32"},"churn":{"type":"string","format":"byte"},"probability":{"type":"number","format":"double"},"timestamp":{"type":"string","format":"date-time"}}},"AdditionalChargesDTO":{"type":"object","properties":{"totalAmount":{"type":"number"},"breakdown":{"type":"array","items":{"$ref":"#/components/schemas/CategoryChargeDTO"}}}},"CategoryChargeDTO":{"type":"object","properties":{"type":{"type":"string"},"amount":{"type":"number"},"percentage":{"type":"number","format":"double"}}},"ClientStatisticsDTO":{"type":"object","properties":{"monthlyAttendanceLastSixMonths":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}},"additionalChargesByCat":{"$ref":"#/components/schemas/AdditionalChargesDTO"}}},"GlobalStatisticsDTO":{"type":"object","properties":{"total":{"type":"integer","format":"int64"},"active":{"type":"integer","format":"int64"},"averageAge":{"type":"number","format":"double"}}}},"securitySchemes":{"bearer-key":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}